src/external/handlers/webservices.ts
import { WebQueryState } from '../../../spec/WebQueries';
import { QueryError } from '../errors';
import { validateUrl } from '../validation';
import { handleError } from './base';

export async function executeRequest(state: WebQueryState, url: string): Promise<any> {
  const query = state.queries.get(url);
  if (!query) {
    throw new QueryError(`Query for URL ${url} not found`);
  }

  try {
    validateUrl(url);
    // Placeholder for actual HTTP request
    // In a real implementation, this would:
    // 1. Build request with headers/params
    // 2. Execute request
    // 3. Handle response
    return {};
  } catch (error) {
    throw handleError(error, QueryError, 'Failed to execute request');
  }
}

export async function validateEndpoint(url: string): Promise<boolean> {
  try {
    validateUrl(url);
    // Placeholder for actual endpoint validation
    // In a real implementation, this would:
    // 1. Send HEAD/OPTIONS request
    // 2. Check response status
    // 3. Validate CORS if needed
    return true;
  } catch (error) {
    throw handleError(error, QueryError, 'Failed to validate endpoint');
  }
}